requirements.txt
requests==2.32.3
pytest==8.3.4
pytest-html==4.1.1
PyYAML==6.0.2
jsonschema==4.23.0
tenacity==9.0.0
pydantic==2.10.4
fastapi==0.115.6
uvicorn==0.34.0
allure-pytest==2.13.5

